requirements.txt
langchain>=0.3.27
langchain-core>=0.3.72
langgraph>=0.6.3
pip>=25.2
pydantic>=2.11.7
python-dotenv>=1.1.1
langchain-openai>=0.2.0
